code/requirements.txt
easydict==1.10
matplotlib==3.5.2
numpy==1.22.3
opencv_contrib_python==4.6.0.66
pandas==1.4.2
Pillow==9.3.0
PyYAML==6.0
scikit_image==0.19.3
scikit_learn==1.2.0
skimage==0.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0